requirements.txt
gym
numpy== 1.24.0
pettingzoo
matplotlib
imageio
inferactively-pymdp
pettingzoo